requirements.txt

embedchain
openai==1.30.2
python-dotenv==1.0.1
streamlit==1.34.0
openpyxl
xlrd
unstructured
networkx